pyproject.toml
[tool.poetry]
name = "graph"
version = "0.1.0"
description = ""
authors = ["Rodrigo Azevedo <rodrigozcma@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
neo4j-driver = "^4.3.6"
neomodel = "^4.0.8"
pydantic = "^1.10.2"
starlette = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
